+++ routes/internal/routes.js
@@ -24,5 +24,8 @@ module.exports = {
   arbimon: [
     require('./arbimon/stream'),
     require('./arbimon/project')
+  ],
+  'ai-hub': [
+    require('./ai-hub')
   ]
 }