+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * **internal:** Add united endpoint for stream_source_file and stream_segment creation ([CE-469](https://jira.rfcx.org/browse/CE-469))
 * **internal:** Call Arbimon recordings endpoint from united endpoint of stream_source_file and stream_segment creation ([CE-469](https://jira.rfcx.org/browse/CE-469))
 * **internal:** Add stream permissions endpoint
+* **internal:** GET /interna/ai-hub/reviews endpoints for getting detections left join with annotations
 
 ### Bug Fixes
 * **core:** update PATCH /classifier endpoint to save previous `deployment_parameters` configuration when given only status change ([CE-344](https://jira.rfcx.org/browse/CE-344))